extract shared escapeLike, renderLikePattern, serializeValue into dialect/utils
