+++ src/lib/orm/dialect/utils.ts
@@ -1,4 +1,46 @@
-import type { Dialect } from "../types.js";
+import type { ColumnKind, Dialect } from "../types.js";
+
+export function escapeLike(s: string) {
+  return s.replaceAll("%", "\\%").replaceAll("_", "\\_");
+}
+
+export function renderLikePattern(
+  mode: "startsWith" | "endsWith" | "contains",
+  value: string,
+) {
+  const escaped = escapeLike(value);
+
+  if (mode === "startsWith") {
+    return `${escaped}%`;
+  }
+
+  if (mode === "endsWith") {
+    return `%${escaped}`;
+  }
+
+  return `%${escaped}%`;
+}
+
+export function serializeSqlValue(kind: ColumnKind, value: unknown) {
+  if (kind === "date") {
+    if (value instanceof Date) {
+      return value.toISOString();
+    }
+
+    return value;
+  }
+
+  if (kind === "json" || kind === "jsonb") {
+    return JSON.stringify(value);
+  }
+
+  if (kind === "boolean") {
+    if (value === true) return 1;
+    if (value === false) return 0;
+  }
+
+  return value;
+}
 
 export function inferDialectFromUrl(url: string): Dialect {
   if (url.startsWith("postgres://") || url.startsWith("postgresql://")) {
